fix: Fix false OK when expected string matches first but is longer
`"stop here -->   "` and `"stop here -->   abcde"` were both `OK`, with the same test function.
In a nutshell, when `read` read everything, it was not compared with the expected output if it should have ended already.
This bug was introduced by the previous stdout comparison fix 94369cc, so it was not present in the eval version.